simple-python-app/requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12"
aiohappyeyeballs==2.6.1 ; python_version >= "3.12"
aiohttp==3.11.12 ; python_version >= "3.12"
aiosignal==1.3.2 ; python_version >= "3.12"
annotated-types==0.7.0 ; python_version >= "3.12"
anyio==4.9.0 ; python_version >= "3.12"
asn1crypto==1.5.1 ; python_version >= "3.12"
attrs==25.3.0 ; python_version >= "3.12"
cachetools==5.5.2 ; python_version >= "3.12"
certifi==2025.1.31 ; python_version >= "3.12"
cffi==1.17.1 ; python_version >= "3.12" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.12"
click==8.1.8 ; python_version >= "3.12"
cloud-sql-python-connector==1.18.0 ; python_version >= "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.12"
cryptography==44.0.2 ; python_version >= "3.12"
dnspython==2.7.0 ; python_version >= "3.12"
email-validator==2.2.0 ; python_version >= "3.12"
fastapi-cli==0.0.7 ; python_version >= "3.12"
fastapi==0.115.12 ; python_version >= "3.12"
frozenlist==1.5.0 ; python_version >= "3.12"
google-api-core==2.24.2 ; python_version >= "3.12"
google-auth==2.38.0 ; python_version >= "3.12"
google-cloud-secret-manager==2.23.2 ; python_version >= "3.12"
googleapis-common-protos==1.69.2 ; python_version >= "3.12"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.12"
grpcio-status==1.71.0 ; python_version >= "3.12"
grpcio==1.71.0 ; python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12"
httpcore==1.0.7 ; python_version >= "3.12"
httptools==0.6.4 ; python_version >= "3.12"
httpx==0.28.1 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
jinja2==3.1.6 ; python_version >= "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.12"
markupsafe==3.0.2 ; python_version >= "3.12"
mdurl==0.1.2 ; python_version >= "3.12"
multidict==6.3.0 ; python_version >= "3.12"
pastel==0.2.1 ; python_version >= "3.12"
pg8000==1.31.2 ; python_version >= "3.12"
poethepoet==0.33.1 ; python_version >= "3.12"
propcache==0.3.1 ; python_version >= "3.12"
proto-plus==1.26.1 ; python_version >= "3.12"
protobuf==5.29.4 ; python_version >= "3.12"
pyasn1-modules==0.4.2 ; python_version >= "3.12"
pyasn1==0.6.1 ; python_version >= "3.12"
pycparser==2.22 ; python_version >= "3.12" and platform_python_implementation != "PyPy"
pydantic-core==2.33.0 ; python_version >= "3.12"
pydantic==2.11.0 ; python_version >= "3.12"
pygments==2.19.1 ; python_version >= "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.12"
python-dotenv==1.1.0 ; python_version >= "3.12"
python-multipart==0.0.20 ; python_version >= "3.12"
pyyaml==6.0.2 ; python_version >= "3.12"
requests==2.32.3 ; python_version >= "3.12"
rich-toolkit==0.14.0 ; python_version >= "3.12"
rich==13.9.4 ; python_version >= "3.12"
rsa==4.2 ; python_version >= "3.12"
scramp==1.4.5 ; python_version >= "3.12"
shellingham==1.5.4 ; python_version >= "3.12"
six==1.17.0 ; python_version >= "3.12"
sniffio==1.3.1 ; python_version >= "3.12"
sqlalchemy==2.0.40 ; python_version >= "3.12"
sqlmodel==0.0.24 ; python_version >= "3.12"
starlette==0.46.1 ; python_version >= "3.12"
typer==0.15.2 ; python_version >= "3.12"
typing-extensions==4.13.0 ; python_version >= "3.12"
typing-inspection==0.4.0 ; python_version >= "3.12"
urllib3==2.3.0 ; python_version >= "3.12"
uvicorn==0.34.0 ; python_version >= "3.12"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.12"
watchfiles==1.0.4 ; python_version >= "3.12"
websockets==15.0.1 ; python_version >= "3.12"
yarl==1.18.3 ; python_version >= "3.12"
